+++ packages/x6/src/research/core/animation.ts
