+++ packages/x6/src/research/core/node.ts
@@ -2,7 +2,7 @@ import { StringExt, ObjectExt, NumberExt } from '../../util'
 import { Point, Rectangle, Angle } from '../../geometry'
 import { Size, KeyValue } from '../../types'
 import { Cell } from './cell'
-import { PortData } from './port-data'
+import { PortData } from './port'
 import { Markup } from './markup'
 import { Edge } from './edge'
 import { Store } from './store'
